% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_genlasso_for_graph.R
\name{admm_genlasso_for_graph}
\alias{admm_genlasso_for_graph}
\title{Generalized Lasso Solver for Graphs Using ADMM}
\usage{
admm_genlasso_for_graph(
  A,
  b,
  D = diag(length(b)),
  C,
  lambda1 = 1,
  lambda2 = 1,
  rho = 1,
  abstol = 1e-04,
  reltol = 0.01,
  maxiter = 1000
)
}
\arguments{
\item{A}{Data matrix.}

\item{b}{Response vector.}

\item{D}{Regularization matrix, defaults to the identity matrix with size equal to the length of \code{b}.}

\item{C}{Constraint matrix for the graph structure.}

\item{lambda1}{Regularization parameter for the lasso penalty.}

\item{lambda2}{Regularization parameter for the graph constraint.}

\item{rho}{Augmentation parameter for ADMM.}

\item{abstol}{Absolute tolerance level for convergence.}

\item{reltol}{Relative tolerance level for convergence.}

\item{maxiter}{Maximum number of iterations for the algorithm.}
}
\value{
A list containing the following components:
\itemize{
\item \code{x}: The solution vector.
\item \code{x_iter}: Solution vector at each iteration.
\item \code{history}: A data frame containing the objective value, primal and dual norms, and the epsilon values for primal and dual convergence checks at each iteration.
}
}
\description{
Solves the generalized lasso problem for graph-structured data using the Alternating Direction Method of Multipliers (ADMM). This function is designed to handle cases where the data matrix has a graph structure.
}
\examples{
# Generate synthetic data
n <- 200; m <- 100; p <- 0.1
A <- matrix(rnorm(m * n), nrow = m)
b <- matrix(rnorm(m))
D <- diag(n)
C <- matrix(rnorm(m * n), nrow = m)

# Solve using admm_genlasso_for_graph
result <- admm_genlasso_for_graph(A, b, D, C)

}
