% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_chr_to_factor.R
\name{all_chr_to_factor}
\alias{all_chr_to_factor}
\title{\strong{\link{!!}} Convert all character variables to factors}
\usage{
all_chr_to_factor(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with character variables converted to factors.
}
\description{
A convenience function to convert all character variables in a data frame
to factor variables.
}
\examples{
library(biostat)
data(PlantGrowth)


# Basic syntax
PlantGrowth <- all_chr_to_factor(PlantGrowth)


# Update the same object using operator from `magrittr` package
library(magrittr)
PlantGrowth \%<>\% all_chr_to_factor()


# Create a data frame
df <- data.frame(
  letters = letters[1:5],
  letters2 = LETTERS[1:5],
  stringsAsFactors = FALSE
)

mapply(class, df) # show classes of columns

# Convert all character variables to strings
df2 <- all_chr_to_factor(df)

# Check the classes in each column
mapply(class, df2)

# # Works with tibbles too
# tbl  <- tibble::as.tibble(df)
# tbl2 <- all_chr_to_factor(tbl)
# mapply(class, tbl2)
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# v0.3.1:
}
