% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_biostat-package.R
\docType{package}
\name{biostat}
\alias{biostat}
\alias{biostat-package}
\title{A Collection of Functions for Biostatistics and Biometry Lectures}
\description{
The package contains a collection of functions dedicated for
biostatistics and biometry lectures (GMF1059) that do some common
statistical routines without writing to much code.
The package is still in its development and usage of some functions is
for demonstration purposes only as they may change in the future.
}
