% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_freq_table_cont.R
\name{bs_freq_table_cont}
\alias{bs_freq_table_cont}
\alias{pander.bs_freq_table_cont}
\title{Frequency table for continuous data}
\usage{
bs_freq_table_cont(y, nbins = NULL, bin_width = NULL, xmin = NULL,
  xmax = NULL, max_bin_decimals = 2, percent_decimals = 1,
  right = TRUE)

\method{pander}{bs_freq_table_cont}(x, ..., caption = "Frequency table",
  justify = c("crrrr"), style = "simple")
}
\arguments{
\item{y}{(numeric) A numeric vector with data.}

\item{nbins}{(integer) Number of bins.}

\item{bin_width}{(number) Bin width.}

\item{xmin}{(number) Lower position of the lowest bin. Not bigger than \code{min(x)}.}

\item{xmax}{(number) Upper position of the highest bin. Not smaller than \code{max(x)}.}

\item{max_bin_decimals}{(integer) Maximum number of significant decimal digits to display in bin names.}

\item{percent_decimals}{(integer) Number of decimal numbers in percentages.}

\item{right}{(logical) Flag that indicates if the intervals should be closed
on the right and open on the left (default) or vice versa.}

\item{x}{object to print}

\item{...}{further arguments to methods}

\item{caption}{(string) caption}

\item{justify}{(string) Column justification}

\item{style}{(string) the name of style for table.}
}
\value{
Object of classes \code{bs_freq_table_cont}, \code{data.frame}.
}
\description{
Frequency table for continuous data
}
\examples{
library(biostat)
library(pander)
bs_freq_table_cont(iris$Sepal.Length, nbins = 5)
bs_freq_table_cont(iris$Sepal.Length, bin_width = 1, xmin = 4)
bs_freq_table_cont(iris$Sepal.Length, nbins = 5) \%>\% pander()
}
