% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mean_boot.R
\name{ci_mean_boot}
\alias{ci_mean_boot}
\title{Basic nonparametric bootstrap confidence interval of mean}
\usage{
ci_mean_boot(y, conf_level = 0.95, repetitions = 2000, na.rm = TRUE,
  resampled_means = FALSE, return_df = TRUE)
}
\arguments{
\item{y}{(numeric) A numeric vector from which \code{NA}s will be removed automatically.}

\item{conf_level}{(number) Confidence level. Number from 0 to 1. Default 0.95.}

\item{repetitions}{(integer) Number of bootstrap resamples.}

\item{na.rm}{(logical) If \code{TRUE} (default), missing values (\code{NA}'s) are removed automatically.}

\item{resampled_means}{(logical) If \code{TRUE}, the vector of bootstrapped means will be returned as the \code{resampled_means} attribute of the returned object.}

\item{return_df}{(logical) If \code{TRUE} (default), result is returned as a data frame. If \code{FALSE} - as a matrix.}
}
\value{
An object (data frame or matrix) with a point estimate and confidence interval of mean.
}
\description{
A very fast implementation of the basic nonparametric bootstrap for obtaining confidence limits for the population mean without assuming normality.
}
\details{
The function is based on code of function \code{\link[Hmisc]{smean.cl.boot}()} from package \pkg{Hmisc}.
}
\examples{
set.seed(111555)
ci_mean_boot(1:60)

set.seed(999555)
y <- rnorm(35, 10, 5)
ci_mean_boot(y)

}
