% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mean_t.R
\name{ci_mean_t}
\alias{ci_mean_t}
\title{t-distribution based confidence interval of mean}
\usage{
ci_mean_t(y, conf_level = 0.95, na.rm = TRUE, return_df = TRUE)
}
\arguments{
\item{y}{(numeric) A numeric vector from which \code{NA}s will be removed automatically.}

\item{conf_level}{(number) Confidence level. Number from 0 to 1. Default 0.95.}

\item{na.rm}{(logical) If \code{TRUE} (default), missing values (\code{NA}'s) are removed automatically.}

\item{return_df}{(logical) If \code{TRUE} (default), result is returned as a data frame. If \code{FALSE} - as a matrix.}
}
\value{
An object (data frame or matrix) with a point estimate and confidence interval of mean.
}
\description{
Student's t-distribution based confidence interval of mean
}
\examples{
set.seed(999555)
y <- rnorm(35, 10, 5)
ci_mean_t(y)
#
}
