% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_standardized.R
\name{coef_standardized}
\alias{coef_standardized}
\alias{print.lm_beta}
\alias{summary.lm_beta}
\alias{print.lm_beta_summary}
\title{\strong{\link{!!}} Compute Standardized Regression Coefficients}
\usage{
coef_standardized(obj)

\method{print}{lm_beta}(x, ..., digits = 3)

\method{summary}{lm_beta}(object, ..., digits = 3)

\method{print}{lm_beta_summary}(x, ..., digits = 3)
}
\arguments{
\item{obj}{(\code{lm} object) A result of function \code{lm()}.}

\item{x}{\code{lm_beta} object.}

\item{...}{further parameters to \code{print} method.}

\item{digits}{(integer) number of decimal places to round the answer to.
Default is 3.}

\item{object}{\code{lm_beta} object.}
}
\value{
Object of class \code{lm_beta} which is a list with 2 named fields:
\itemize{
\item \code{b} named numeric vector with regression coefficients (not standardized);
\item \code{beta} named numeric vector  with standardized coefficients from \code{lm()} model.
}
}
\description{
Compute the standardized regression coefficients (beta) from an object of class \code{lm}).
}
\details{
This function is inspired by function  \pkg{QuantPsyc}\verb{::lm.beta()} written by Thomas D. Fletcher. \cr
\code{coef_standardized()} provides standardized coefficients even when interaction members are present. This is achieved by computing whole model matrix (with all right-hand side members of formula used in call of \code{lm()}) and calculating standard deviations of each regressor (including interaction members) based on these columns. \cr
\code{coef_standardized()} does not fail if intercept is not present.\cr

The remaining calculations are the same as in \pkg{QuantPsyc}\verb{::lm.beta()}.
}
\examples{

data(USJudgeRatings)
us <- USJudgeRatings
names(us)

lm1 <- lm(CONT ~ INTG + DMNR + log(DILG), data = us)
coef_standardized(lm1)

lm2 <- lm(CONT ~ INTG + DMNR * DILG, data = us)
coef_standardized(lm2)

summary(coef_standardized(lm2))


# Do not include intercept
lm3 <- lm(CONT ~ 0 + INTG, data = us)
coef_standardized(lm3)
summary(coef_standardized(lm3))
}
\seealso{
\itemize{
\item \code{\link[stats:lm]{stats::lm()}} in package \pkg{stats}.
\item \code{\link[QuantPsyc:lm.beta]{QuantPsyc::lm.beta()}} in package \pkg{QuantPsyc}.
}
}
\author{
Vilmantas Gegzna (modified function written by Thomas D. Fletcher).
}
\keyword{models}
