% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_standardized.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{standardized_coef}
\title{Deprecated functions in `BioStat` package}
\usage{
standardized_coef(obj)
}
\arguments{
\item{obj}{\code{lm} object.}
}
\description{
Depreceted functions. They exist for compatibility with
             previous versions.
}
