% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_summary.R
\name{do_summary}
\alias{do_summary}
\alias{print.num_summaries}
\title{Do numerical summaries by groups}
\usage{
do_summary(y, data = NULL, stat = c("n", "missing", "mean", "sd", "var",
  "min", "Q1", "Md", "Q3", "max", "IQR", "skewness", "kurtosis"))

\method{print}{num_summaries}(x, ..., digits = NA, format = "f")
}
\arguments{
\item{y}{formula with variable names to summarize. See more in examples.}

\item{data}{data set}

\item{stat}{(character) Descriptive statistics to compute. Currently supported statistics:
\itemize{
    \item \code{"n"} - number of non-missing observations,
    \item \code{"missing"} - number of missing observations,,
    \item \code{"mean"} - arithmetic mean,
    \item \code{"sd"} - standard deviation,
    \item \code{"var"} - variance,
    \item \code{"min"} - minimum value,
    \item \code{"Q1"} - 1-st quartile,
    \item \code{"Md"} - median,
    \item \code{"Q3"} - 3-rd quartile,
    \item \code{"max"} - maximum value,
    \item \code{"IQR"} - interquartile range,
    \item \code{"skewness"} - skewness,
    \item \code{"kurtosis"} - excess kurtosis.
}}

\item{x}{object to print}

\item{...}{further arguments to methods.}

\item{digits}{(numeric or \code{NA}) a desired number of digits after the decimal point (if \code{format = "f"}) or a number of significant digits (\code{format = "g"}, \code{= "e"} or \code{= "fg"}). \cr
Either one integer (to set the same formatting for all columns) or a vector of inegers (to set formatting for each column separately).\cr
Use \code{NA} to leave column unformatted.}

\item{format}{(character) \code{"f"}, \code{"g"}, \code{"e"}, \code{"fg"}. Either one value or a vector of values for each column.
Each value will be passed to \code{fun} separately.\cr
\code{"f"} gives numbers in the usual \code{xxx.xxx} format;\cr
\code{"e"} and \code{"E"} give \code{n.ddde+nn} or \code{n.dddE+nn} (scientific format);\cr
\code{"g"} and \code{"G"} put number into scientific format only if it saves space to do so.\cr
\code{"fg"} uses fixed format as \code{"f"}, but digits as the minimum number of significant digits. This can lead to quite long result strings}
}
\value{
Data frame with summary satatistics.
}
\description{
Do numerical summaries by groups with formaula interface. Missing values are automatically removed.
}
\examples{
library(BioStat)
data(cabbages, package = "MASS")

do_summary(~VitC, data = cabbages) \%>\%
    print(digits = 2)

do_summary(VitC ~ Cult, data = cabbages) \%>\%
    print(digits = 2)

do_summary(VitC ~ Cult + Date, data = cabbages, stat = "mean") \%>\%
    print(digits = 2)

do_summary(HeadWt + VitC ~ Cult + Date,
           data = cabbages,
           stat = c("n", "mean", "sd")) \%>\%
    print(digits = 1)

}
