% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a0_pagalbines_funkcijos.R
\name{eval_}
\alias{eval_}
\title{eval text as a command}
\usage{
eval_(x, envir = parent.frame(), ...)
}
\arguments{
\item{x}{(string) A text with command to evaluate.}

\item{envir}{An environment to evaluate in.}

\item{...}{further parameters to \code{\link[base:eval]{base::eval()}}.}
}
\description{
eval text as a command
}
\keyword{internal}
