% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.R
\name{format_p_values}
\alias{format_p_values}
\alias{format_p}
\alias{format_as_p_columns}
\alias{add_signif_stars}
\alias{get_signif_stars}
\alias{signif_stars_legend}
\alias{rm_zero}
\title{Format p-values}
\usage{
format_p_values(p, digits_p = 3, signif_stars = TRUE, rm_zero = FALSE,
  add_p = FALSE, rm_spaces = FALSE)

format_p(p_i, digits_p = 3, signif_stars = TRUE, rm_zero = FALSE,
  add_p = FALSE, rm_spaces = FALSE)

format_as_p_columns(data, colnames = c("p.value", "p.adjust"), digits_p = 3,
  rm_zero = FALSE, signif_stars = FALSE, ...)

add_signif_stars(p)

get_signif_stars(p)

signif_stars_legend()

rm_zero(str)
}
\arguments{
\item{p}{A (vector of) p-value(s). Numeric or coercible to numeric.}

\item{digits_p}{(numeric) Number of significant digits to round a p value to.
                  No less than 2. \itemize{
\item if \code{digits_p = 2}: \enumerate{
      \item values below 0.001 are printed as \code{"<0.001"};
      \item values between 0.001 and 0.01 as \code{"<0.01"};
      \item all other values are rounded to two decimal places.
      }
\item if \code{digits_p = 3}, only formatting \code{"<0.01"} is skipped.
}}

\item{signif_stars}{(logical) Flag if significance stars should be added to each p value.
Not less than 2.}

\item{rm_zero}{(logical) Flag if leading zero before the point should be
removed.}

\item{add_p}{(logical) Flag if letter "p" should included in the expression.}

\item{rm_spaces}{(logical) Flag if all spaces should be removed.}

\item{p_i}{(numeric) A single p value.}

\item{data}{A dataset.}

\item{colnames}{(character) vector with column names to be formatted as p values.}

\item{...}{Arguments to further methods.}

\item{str}{A string or convertible to string}
}
\value{
A character vector with formatted p values.
}
\description{
Fuctions to fromat p values.
}
\details{
\itemize{
 \item \code{format_p} - formats single p value.
 \item \code{format_as_p_columns} - formats indicated numeric columns in a dataframe as p values (columns are converted into strings).
 \item \code{get_signif_stars} -  takes numeric p values brings appopriate stars of statistical significance.
 \item \code{add_signif_stars} - formats numeric p values by adding significance stars (result is character vector).
 \item \code{signif_stars_legend} - generates legend for significance stars (result is a string).
 \item \code{rm_zero} - function removes zero at the beginning of a number (returns a  string with the same value but without the leading zero).
}
}
\examples{
library(BioStat)

# Prettify p-values

format_p_values(0.0005)

format_p_values(0.005)
format_p_values(0.005, signif_stars = FALSE)
format_p_values(0.005, rm_zero = TRUE)
format_p_values(0.005, digits_p = 2)
format_p_values(0.005, digits_p = 2, rm_zero = TRUE, signif_stars = FALSE)

format_p_values(0.00022)
format_p_values("0.00022")
format_p_values("0.052")

format_p_values(c(0.005, 0.0005, 0.052147))


get_signif_stars(0.005)
add_signif_stars(0.005)

get_signif_stars(0.0005)
add_signif_stars(0.0005)

get_signif_stars(0.052147)
add_signif_stars(0.052147)

signif_stars_legend()


format_p(.02, digits_p = 2)
format_p(.0002)
format_p(.0002, signif_stars = FALSE)

rm_zero(0.02)

}
