% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unique}
\alias{is_unique}
\alias{is_not_inique}
\title{Test which values are/are not unique in a vector}
\usage{
is_unique(x)

is_not_inique(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Vector of logical values indicating \code{TRUE} for values which are unique.
}
\description{
Test which values are/are not unique in a vector
}
\examples{

is_unique(1:10)

is_unique(c(1, 1, 1))

x <- c(1, 2, 3, 4, 3, 3, 1)
rez <- is_unique(x)
rez

data.frame(x = x, is_unique = rez)
}
\keyword{utilities}
