% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_class.R
\name{ls_class}
\alias{ls_class}
\title{List objects of a certain class.}
\usage{
ls_class(class = NULL, all.names = FALSE, envir = parent.frame())
}
\arguments{
\item{class}{(character) The name of class.}

\item{all.names}{(logical) If \code{TRUE}, all object names are returned.
If \code{FALSE}, names which begin with a \code{.} are omitted.}

\item{envir}{The environment to search for objects in.}
}
\value{
Character vector.
}
\description{
List objects of a certain class.
}
