% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cld.R
\name{make_cld}
\alias{make_cld}
\alias{make_cld.pairwise.htest}
\alias{make_cld.posthocTGH}
\alias{make_cld.posthoc_anova}
\alias{make_cld.PMCMR}
\alias{make_cld.formula}
\alias{make_cld.matrix}
\alias{make_cld.data.frame}
\alias{make_cld.pairwise_pval_df}
\title{Make a compact letter display (cld) for pair-wise comparison}
\usage{
make_cld(obj, ..., alpha = 0.05)

\method{make_cld}{pairwise.htest}(obj, ..., alpha = 0.05)

\method{make_cld}{posthocTGH}(obj, ..., alpha = obj$intermediate$alpha)

\method{make_cld}{posthoc_anova}(obj, ..., alpha = 1 - obj$input$conf_level)

\method{make_cld}{PMCMR}(obj, ..., alpha = 0.05)

\method{make_cld}{formula}(obj, ..., data = NULL, alpha = 0.05)

\method{make_cld}{matrix}(obj, ..., alpha = 0.05)

\method{make_cld}{data.frame}(obj, ..., formula = p.adjust ~ Comparison,
  alpha = 0.05)

\method{make_cld}{pairwise_pval_df}(obj, ..., alpha = 0.05)
}
\arguments{
\item{obj}{Object with pair-wise comparisons (e.g., post-hoc test results).
Currently supported objects: \itemize{
\item \emph{posthocTGH} from package \pkg{userfriendlyscience};
\item \emph{PMCMR} from package \pkg{PMCMR}.
}}

\item{...}{Further arguments to methods.}

\item{alpha}{(numeric from 0 to 1) Significance level.}

\item{data}{A dataset with p values and names of comparisons. This argument
is used if \code{obj} is formula. More details in examples.}

\item{formula}{An R model \code{\link[stats]{formula}} where left-hand side
term indicates variable with p values and right-hand side term defines
variable with comparisons, e.g. \code{p.adjust ~ Comparison}. Usually is
used in combination with \code{data}.}
}
\value{
(A data frame with) compact letter display.
}
\description{
Make a compact letter display for results of pair-wise comparisons
(e.g., ANOVA post-hoc tests, Kruskal-Wallis post-hoc tests and other).
}
\examples{
library(biostat)

# Example 1: class `pairwise.htest`

obj1 <- pairwise.wilcox.test(chickwts$weight, chickwts$feed, exact = FALSE)
make_cld(obj1)


# Example 2: class `pairwise.htest`

obj2 <- with(OrchardSprays, pairwise.t.test(decrease, treatment))
make_cld(obj2)


# Example 3: class `pairwise.htest`

\donttest{
smokers  <- c(83, 90, 129, 70)
patients <- c(86, 93, 136, 82)
obj3 <- pairwise.prop.test(smokers, patients)

make_cld(obj3)
}

# Example 4: class `PMCMR`

\donttest{
obj4 <- PMCMR::posthoc.kruskal.conover.test(count ~ spray,
                                            data = InsectSprays)
make_cld(obj4)
}


# Example 5: class `posthocTGH`

obj5 <- posthoc_anova(weight ~ Diet,
                      data = ChickWeight,
                      method = "Games-Howell")
make_cld(obj5)


# Example 6: class `posthoc_anova`

obj6 <- posthoc_anova(weight ~ Diet,
                     data = ChickWeight,
                     method = "Games-Howell")
make_cld(obj6)


# Example 7: class `formula`

my_dataframe <- data.table::fread(
    'Comparison     p.value p.adjust
    "EE - GB = 0"        1 1.000000
    "EE - CY = 0" 0.001093 0.003279
    "GB - CY = 0" 0.005477 0.008216')

make_cld(p.adjust ~ Comparison, data = my_dataframe)


# Example 8: class `matrix`

# (for symetric matrices of p values)

# Create matrix
m <- c(1.00, 0.22, 0.05, 0.00,
       0.22, 1.00, 0.17, 0.01,
       0.05, 0.17, 1.00, 0.22,
       0.00, 0.01, 0.22, 1.00)
obj8 <- matrix(m, nrow = 4)
rownames(obj8) <- colnames(obj8) <- c("P", "O", "I", "U")
obj8

# Make cld
make_cld(obj8)


}
