% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cld_df.R
\name{make_cld_df}
\alias{make_cld_df}
\title{[!] Make a data frame with cld}
\usage{
make_cld_df(formula = NULL, data = NULL, comparison = NULL,
  p.value = NULL, threshold = 0.05, print.comp = FALSE,
  remove.space = TRUE, remove.equal = TRUE, swap.colon = TRUE,
  swap.vs = FALSE, reversed = FALSE, swap_compared_names = FALSE, ...)
}
\arguments{
\item{formula}{The formula used to make the test (lm, aov, glm, etc.).  Like
y ~ x.}

\item{data}{Data used to make the test.}

\item{threshold}{Second (reference) argument to "compare".}

\item{reversed}{A logical value indicating whether the order of the letters
should be reversed. Defaults to FALSE.}

\item{swap_compared_names}{(logical) \cr If \code{TRUE}, group names are swapped
from, e.g., "2-1" or "second-first" to "1-2" or "first-second". This leads to
different order of cld letters and different order of compared groups (if
without swapping the result is incorrect order of groups.)}

\item{...}{Extra arguments passed to multcompLetters.}
}
\description{
Compute a compact letter display (cld) of statistically significant
differences in pairwise comparisons and output he results as a dataframe.
}
\note{
This function is based on code in function \code{\link[rcompanion]{cldList}()}
from package \pkg{rcompanion} by Salvatore Mangiafico.
}
\keyword{internal}
