% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{make_relative_path}
\alias{make_relative_path}
\title{\strong{\link{!!!}}}
\usage{
make_relative_path(path, wd = getwd(), warn = TRUE)
}
\arguments{
\item{path}{string with absolute path.}

\item{wd}{string of path to absolute path.}

\item{warn}{(logical) flag to warn if folder does not exist on a computer.}
}
\value{
string
}
\description{
\strong{\link{!!!}}
}
\examples{
# No examples yet
}
