% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{min_numeric}
\alias{min_numeric}
\title{Min for numeric vectors or NA othervise}
\usage{
min_numeric(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical passed to min}
}
\value{
Return minimum value if \code{x} is numeric and \code{NA} oterwise.
}
\description{
Min for numeric vectors or NA othervise
}
\keyword{internal}
