% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse model formula.}
\usage{
parse_formula(formula, data = NULL, keep_all_vars = FALSE)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} of form: \itemize{
     \item \code{~ y};
     \item \code{~ y1 + y2};
     \item \code{y ~ x};
     \item \code{y1 + y2 ~ x1 + x2};
     \item \code{~ y|condition};
     \item \code{y ~ x|condition};
     \item \code{y1 + y2 ~ x1 + x2|cond1 + cond2};
     \item  or similar.
}
Variables can be expressed as function, e.g., \code{log(x)}, \code{as.factor(x)}, which will be evaluated. Dot (\code{.}) is not accepted.}

\item{data}{A data frame.}

\item{keep_all_vars}{(logical) Flag indicating if all variables
(even those not in formula) should be included in the
output data.
Expression \code{keep_all_vars = TRUE} has effect
only if \code{data} is a data frame (and not \code{NULL}
or an environment).}
}
\value{
A list with fields: \enumerate{
    \item \code{formula} - the formula used;
    \item \code{names} - alist of character vectors with names in formula arranged in certain way (see below);
    \item \code{data} - a data frame with variables and evaluated expressions in the formula and, optionally, with the remaining variables from the original data frame.
}

Contents of the field \code{names}: \itemize{
    \item \code{all_names} - all names (and expressions) in the formula;
    \item \code{y}, \code{x} - names for "x" and "y" variables (except condition variables);
    \item \code{lhs}, \code{rhs} - names in left-hand and right-hand side of formula (except condition variables);
    \item \code{condition} - names of condition variables and expressions;
    \item \code{gr} - if condition exists, \code{gr} is the same as condition, otherwise it is the same as \code{x}.
}
}
\description{
Parse model formula, create new variables by evaluating expressions and return
parsed names and new data frame. These elements are ment to be used by other
functions.
}
