% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_anova.R
\name{posthoc_anova}
\alias{posthoc_anova}
\alias{posthoc_anova.formula}
\alias{posthoc_anova.default}
\alias{print.posthoc_anova}
\alias{plot.posthoc_anova}
\title{[!] Post-hoc tests for ANOVA and Welch ANOVA}
\usage{
posthoc_anova(y, group = NULL, method = c("Games-Howell", "Tukey"),
  conf_level = 0.95, digits = 3, digits_p = 3, p_adjust = "none",
  format_pvalue = TRUE, data = NULL, ...)

\method{posthoc_anova}{formula}(y, group = NULL,
  method = c("Games-Howell", "Tukey"), conf_level = 0.95, digits = 3,
  digits_p = 3, p_adjust = "none", format_pvalue = TRUE,
  data = NULL, ..., sep = " | ")

\method{posthoc_anova}{default}(y, group = NULL,
  method = c("Games-Howell", "Tukey"), conf_level = 0.95, digits = 3,
  digits_p = 3, p_adjust = "none", format_pvalue = TRUE,
  data = NULL, ...)

\method{print}{posthoc_anova}(x, digits = x$input$digits,
  digits_p = x$input$digits_p, digits_param = 2, ...)

\method{plot}{posthoc_anova}(x, ..., zero_line_color = "grey",
  add_p = TRUE, p_size = 1, p_color = "blue", p_pos_adj = 0.22,
  flip_xy = TRUE)
}
\arguments{
\item{y}{(numeric|formula)\cr Either a numeric vector or formula.}

\item{group}{(factor)\cr a vector that either is a factor or can be converted into
one. If \code{y} is a formula, \code{group} is ignored.}

\item{method}{(\code{"games-howell"}|\code{"tukey"})\cr Name of post-hoc
tests to conduct. Valid values are "tukey" and "games-howell".}

\item{conf_level}{(number)\cr Confidence level (equals to \code{1 - alpha}, where alpha is significanve level). Number from 0 to 1. Default is 0.95.}

\item{digits}{(integer)\cr The number of digits to round data related numbers to.}

\item{digits_p}{(integer)\cr The number of digits to round p values to. Must be 2, 3 or more.}

\item{p_adjust}{Any valid \code{\link[stats]{p.adjust}} method.}

\item{format_pvalue}{(does not work yet) \cr Whether to format the p values according to APA
standards (i.e. replace all values lower than .001 with '<.001'). This only
applies to the printing of the object, not to the way the p values are
stored in the object.}

\item{data}{(data frame)\cr a data frame with data to be used in combination with formula.}

\item{...}{Further arguments to methods.}

\item{sep}{(character)\cr A string with a symbol to separate group names if several grouping variables are used.}

\item{x}{object to print.}

\item{zero_line_color}{(character) Color for line indicating zero differences.}

\item{add_p}{(logical) Flag if p values should be added.}

\item{p_size}{(numeric) Font size to p-values-related text.}

\item{p_color}{(character) Color for p values.}

\item{p_pos_adj}{(numeric) Factor for p value position correction}

\item{flip_xy}{(logical) Flag if x and y axes should be swapped.}
}
\value{
A list of three elements:
\item{input}{List with input arguments}
\item{output}{List with post-hoc test results}.
}
\description{
Games-Howell, Tukey HSD and other post-hoc tests for ANOVA and Welch ANOVA.
Either Games-Howell test or Tukey honestly significant difference (HSD)
post-hoc tests for one-way analysis of variance (ANOVA).
}
\details{
The main part of the  function code and descriptions were
imported from \code{posthocTGH()} in package \pkg{userfriendlyscience}.
}
\note{
Options that carry out Games-Howell and Tukey HSD analyses are based
on code of function \code{posthocTGH()} in package \pkg{userfriendlyscience}
(version 0.7.0).
}
\examples{
library(biostat)

# Compute post-hoc statistics using the Games-Howell method
posthoc_anova(weight ~ Diet, data = ChickWeight, method = "Games-Howell")

# Compute post-hoc statistics using the Tukey method
posthoc_anova(weight ~ Diet, data = ChickWeight, method = "Tukey")


}
\seealso{
\itemize{
  \item{\url{http://www.gcf.dkf.unibe.ch/BCB/files/BCB_10Jan12_Alexander.pdf}}
  \item{\url{https://rpubs.com/aaronsc32/games-howell-test}}
  \item{\url{http://aoki2.si.gunma-u.ac.jp/R/src/tukey.R}}
  \item{\url{https://stats.stackexchange.com/questions/83941/games-howell-post-hoc-test-in-r}}

}
}
\author{
Gjalt-Jorn Peters (Open University of the Netherlands) \cr
Jeff Bagget (University of Wisconsin - La Crosse) \cr
Vilmantas Gegzna
}
\keyword{utilities}
