% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_line_coeffs.R
\name{qq_line_coeffs}
\alias{qq_line_coeffs}
\title{Compute a slope and an intercept for a reference line in a qq-plot}
\usage{
qq_line_coeffs(
  y,
  datax = FALSE,
  distribution = qnorm,
  probs = c(0.25, 0.75),
  qtype = 7,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{y}{(numeric) A numeric vector.}

\item{datax}{logical. Should data values be on the x-axis?}

\item{distribution}{quantile function for reference theoretical distribution.}

\item{probs}{numeric vector of length two, representing probabilities.
    Corresponding quantile pairs define the line drawn.}

\item{qtype}{the \code{type} of quantile computation used in \code{\link[stats]{quantile}}.}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{...}{Further arguments for \code{distribution}.}
}
\value{
A vector with a \code{slope} and an \code{intercept} for a qq-line.
}
\description{
Compute a slope and an intercept for a reference line ("qq-line")
in a quantile comparison plot given a vector of data an a theoretical
distribution.
}
\details{
The code is based on code of \code{qqline()} in package \pkg{stats}.
But instead of plotting a reference line, it returns its coefficients.
}
\examples{

set.seed(254)
x <- rnorm(50)

qq_line_coeffs(x)

}
\keyword{internal}
