% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{[!] A QQ-plot for multiple groups}
\usage{
qq_plot(y, data = NULL, distribution = "norm", ..., line = c("quartiles",
  "robust", "int=0,slope=1", "0,1", "none"), envelope = 0.95,
  method = c("mle-normal", "trimmed-normal", "moment-normal", "any"),
  labels = NULL, groups = NULL, use_colors = FALSE, scales = "free",
  sep = " | ")
}
\arguments{
\item{y}{(formula|numeric|character)
Either a formula, a numeric vector or a name of a vector
in \code{data}.
If \code{y} is a formula (e.g. \code{variable ~ factor}), left-hand
side provides variable to be summarized. Right-hand side and condition
describe subsets. If the left-hand side is empty, right-hand side and
condition are shifted over as a convenience.}

\item{data}{A data frame that contains the variables mentioned in \code{y}.}

\item{distribution}{root name of comparison distribution -- e.g., \code{"norm"} for the
    normal distribution; \code{t} for the t-distribution.}

\item{...}{Parameters to be passed to function, selected in \code{distribution}.
In \code{print} method, further parameters to function \code{print}.}

\item{line}{(string) A parameter, that controls how reference line is drawn.
           Options:\itemize{
    \item{\code{"0,1"} or \code{"int=0,slope=1"} to plot a line
            with intercept = 0 and slope = 1;}
    \item{ \code{"quartiles"} to pass a line through the quartile-pairs;}
    \item{ \code{"robust"} for a robust-regression line;
            the latter uses the \code{rlm} function from the \pkg{MASS} package};
    \item{option \code{"none"} is not implemented yet.}
}}

\item{envelope}{(numeric | \code{FALSE}) confidence level for point-wise
confidence envelope.}

\item{method}{(string: \code{"trimmed-normal"}, \code{"normal"},
             \code{"any"}). A method, that controls how estimates of
             parameters for \code{distribution} are computed.
     Options:
     \itemize{

           \item{\code{"mle-normal"} (default) all data values are used to
           estimate parameters of theoretical normal distribution using
           method of maximum likelihood;}

           \item{\code{"trimmed-normal"} 10\% of the most extreme
           data values are trimmed before parameters of theoretical normal
           distribution are estimated using method of moments;}

           \item{\code{"moment-normal"} all data values are used to estimate
           parameters of theoretical normal distribution using method of moments;}

           \item{\code{"any"} either parameters are provided manually by user
           or default parameters are used (if any).}
     }
          Options \code{"mle-normal"}, \code{"trimmed-normal"} and
          \code{"moment-normal"} are applicable only if
          \code{distribution = "norm"}.
          Otherwise \code{"any"} is selected automatically.}

\item{groups}{(\code{NULL}|factor|character) \cr
An alternative way to provide groups.
If \code{y} is a numeric vector, \code{groups} must be
a factor (or \code{NULL}). If \code{y} is a sting,
\code{groups} must also be a string (or \code{NULL}).}

\item{use_colors}{(logical) use colors for multiple groups}

\item{scales}{("free"|"free_x"|"free_y"|"fixed")
a parmeter to be passed to
 \code{\link[ggplot2]{facet_wrap}}.}

\item{sep}{(character)
Group name separator if more than one grouping variable is used.
default is \code{"|"}.}
}
\value{
A \code{ggplot2} object
}
\description{
Make a quantile comparison plot (qq-plot) for each subset of groups separately
using \pkg{ggplot2} graphics.
}
\details{
Function \code{qq_plot} is inspired by \code{qqPlot()} in package \pkg{car}
(writen by J. Fox).
}
\examples{
library(biostat)
data(iris)

# Formula (no groups):
qq_plot(~Sepal.Length, data = iris)
qq_plot("Sepal.Length", data = iris)

# Formula (several groups):
qq_plot(Sepal.Length ~ Species, data = iris)

qq_plot(Sepal.Length ~ Species, data = iris, envelope = 0.90)

# Formula (several groups in colors):
qq_plot(Sepal.Length ~ Species, data = iris, use_colors = TRUE)

# Vectors (several groups):
qq_plot(iris$Sepal.Length, groups = iris$Species)

# For one group:
qq_plot("Sepal.Length", data = iris)

qq_plot(~Sepal.Length, data = iris)

qq_plot(iris$Sepal.Length)


# Other examples
qq_plot(~weight, data = chickwts)

qq_plot(weight ~ feed, data = chickwts)

qq_plot(uptake ~ Type + Treatment, data = CO2)
}
\seealso{
\code{\link[car]{qqPlot}} from \pkg{car} package,
         \code{\link[stats]{qqplot}} from \pkg{stats} package.
}
\keyword{ggplot2}
\keyword{plots}
