% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a0_pagalbines_funkcijos.R
\name{rbind_df_in_list}
\alias{rbind_df_in_list}
\title{Row-bind data frames which are inside a list and keep the names of the list fields}
\usage{
rbind_df_in_list(x)
}
\arguments{
\item{x}{A named list of data frames (all data frames must have the same structure.)}
}
\value{
A data frame with field names of input list as a first column (called '.group').
}
\description{
Row-bind data frames which are inside a list and keep the names of the list fields
}
\examples{
# Data:

data <- list(
    A = read.table(text =
"           intercept    slope      type conf
qq_refline -39.55714 1.123535 quartiles 0.95"),

    B = read.table(text =
"           intercept     slope      type conf
qq_refline  29.42192 0.7940267 quartiles 0.95")
)


rbind_df_in_list(data)

# .group intercept     slope      type
# 1      A -39.55714 1.1235350 quartiles
# 2      B  29.42192 0.7940267 quartiles

}
\keyword{internal}
