% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vector.R
\name{scale_vector}
\alias{scale_vector}
\title{Scale a vector}
\usage{
scale_vector(y, center = mean, scale = sd)
}
\arguments{
\item{y}{A numeric vector.}

\item{center}{Either a function that computes center of data
(such as \code{mean}) or a single numeric value.}

\item{scale}{Either a function that computes variability of data
(such as \code{sd}) or a single numeric value.}
}
\value{
The same object as \code{y} just with every element scaled
}
\description{
From every element in a vector, subtract \code{center} and
divide by \code{scale}.
}
\examples{

y <- 1:10
scale_vector(y)

scale_vector(y, center = median, scale = IQR)

scale_vector(y, center = 10, scale = 2)
}
