% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{standardize_path}
\alias{standardize_path}
\title{Standardize string with path}
\usage{
standardize_path(path)
}
\arguments{
\item{path}{string with path}
}
\value{
string ...
}
\description{
Converts "~" into absolute path to home directory, and replaces escaped back
slashes "\\" with forward slash "/".
}
\examples{
standardize_path("D:\\\\folder")
standardize_path("~")
}
