% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_normal_density.R
\name{stat_normal_density}
\alias{stat_normal_density}
\title{A ggplot2 layer that fits and plots normal density function}
\usage{
stat_normal_density(mapping = NULL, data = NULL, geom = "line",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, trim = 0, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{trim}{(numeric) Fraction of points to trim before computing mean and
standard deviation, which are used to construct
theoretical normal distribution.
Number between 0 and 1. Default is \code{0}.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
A ggplot2 layer that fits and plots a theoretical normal density function.
May be used to visually compare between empirical and theoretical normal
distributions.
}
\section{Computed variables}{


\describe{
   \item{density}{normal density}
   \item{ndensity}{normal density scaled to maximum of 1}
   \item{count}{normal density * number of points}
}
}

\examples{
library(ggplot2)
library(BioStat)

ggplot(mtcars, aes(mpg)) + stat_normal_density()

ggplot(mtcars, aes(mpg)) +
     stat_density(alpha = 0.8) +
     stat_normal_density(color = "red", size = 1)

}
\keyword{ggplot2}
\keyword{plots}
