% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{summary_funs}
\alias{summary_funs}
\alias{q1}
\alias{q3}
\alias{n_missing}
\alias{n_ok}
\alias{n_nonmissing}
\alias{iqv}
\alias{iqv.default}
\alias{iqv.table}
\title{\strong{\link{!!}} Descriptive statistics}
\usage{
q1(y, na.rm = TRUE, type = 7, names = FALSE, ...)

q3(y, na.rm = TRUE, type = 7, names = FALSE, ...)

n_missing(y)

n_ok(y)

n_nonmissing(y)

iqv(y, ...)

\method{iqv}{default}(y, ...)

\method{iqv}{table}(y, ...)
}
\arguments{
\item{y}{Variable to summarize}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{type}{an integer between 1 and 9 selecting one of the
    nine quantile algorithms detailed below to be used.}

\item{names}{logical; if true, the result has a \code{\link{names}}
    attribute.  Set to \code{FALSE} for speedup with many \code{probs}.}

\item{...}{futher arguments to methods}
}
\value{
Summary statistic(s).
}
\description{
\strong{\link{!!}} Descriptive statistics
}
\keyword{utilities}
