% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ggedit_code.R
\name{tidy_ggedit_code}
\alias{tidy_ggedit_code}
\title{Extract and tidy code from ggedit object.}
\usage{
tidy_ggedit_code(x, ...)
}
\arguments{
\item{x}{A \code{ggedit} object.}

\item{...}{Other parameters to \code{\link[styler:style_text]{styler::style_text()}}.}
}
\value{
String.
}
\description{
Extract and tidy code from ggedit object.
}
