% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim certain proportion of the most extreme values}
\usage{
trim(y, trim = 0.1, na.rm = FALSE, ...)
}
\arguments{
\item{y}{A numeric vector}

\item{trim}{(number) A proportion of values to be removed.
A number between 0 and 1. Default is 0.1.}

\item{na.rm}{(logical) if \code{TRUE}, any \code{NA} and \code{NaN}'s are
removed from \code{y} before quantiles are computed.}

\item{...}{further argument to \code{\link[stats]{quantile}}.}
}
\value{
A trimmed vector.
}
\description{
Trim certain proportion of the most extreme values in a numeric vector.
}
\examples{
1:10
# [1] 1 2 3 4 5 6 7 8 9 10

trim(1:10)
# [1] 2 3 4 5 6 7 8 9

trim(1:10, trim = 0.4)
# [1] 3 4 5 6 7 8

}
