% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_summary.R
\name{do_summary}
\alias{do_summary}
\alias{print.num_summaries}
\title{Do numerical summaries by groups}
\usage{
do_summary(y, data = NULL, stat = c("n", "missing", "mean", "trimmed", "sd",
  "variance", "min", "Q1", "median", "Q3", "max", "mad", "IQR", "range", "cv",
  "se", "skewness", "kurtosis"), trim = 0.1, type = 3, na.rm = TRUE)

\method{print}{num_summaries}(x, ..., digits = NA, format = "f",
  digits_sk = 2)
}
\arguments{
\item{y}{formula with variable names to summarize. See more in examples.}

\item{data}{data set}

\item{stat}{(character) Descriptive statistics to compute. Currently supported statistics:
\itemize{
    \item \code{"n"} - number of non-missing observations,
    \item \code{"missing"} - number of missing observations,,
    \item \code{"mean"} - arithmetic mean,
    \item \code{"sd"} - standard deviation,
    \item \code{"variance"} - variance,
    \item \code{"trimmed"} - trimmed mean,
    \item \code{"min"} - minimum value,
    \item \code{"Q1"} - 1-st quartile,
    \item \code{"Md"} - median,
    \item \code{"Q3"} - 3-rd quartile,
    \item \code{"max"} - maximum value,
    \item \code{"mad"} - median absolute deviation from median (more details \link[stats]{mad}),
    \item \code{"IQR"} - interquartile range,
    \item \code{"range"} - range,
    \item \code{"cv"} - coefficient of variation,
    \item \code{"se"} - standard error of mean,
    \item \code{"skewness"} - skewness,
    \item \code{"kurtosis"} - excess kurtosis.
}}

\item{trim}{The fraction (0 to 0.5) of observations to be trimmed from each end of sorted variable before the mean is computed. Values of trim outside that range are taken as the nearest endpoint.}

\item{type}{(integer: 1, 2, 3) The type of skewness and kurtosis estimate.
See \code{\link[psych]{describe}} and \code{\link[psych]{mardia}}
for details.}

\item{na.rm}{(logical) Flag to remove missing values. Default is \code{TRUE}.}

\item{x}{object to print}

\item{...}{further arguments to methods.}

\item{digits}{Number of digits for descriptive statistics.}

\item{format}{(character) \code{"f"}, \code{"g"}, \code{"e"}, \code{"fg"}. Either one value or a vector of values for each column.
Each value will be passed to \code{fun} separately.\cr
\code{"f"} gives numbers in the usual \code{xxx.xxx} format;\cr
\code{"e"} and \code{"E"} give \code{n.ddde+nn} or \code{n.dddE+nn} (scientific format);\cr
\code{"g"} and \code{"G"} put number into scientific format only if it saves space to do so.\cr
\code{"fg"} uses fixed format as \code{"f"}, but digits as the minimum number of significant digits. This can lead to quite long result strings}

\item{digits_sk}{Number of digits for skweness and kurtosis.}
}
\value{
Data frame with summary satatistics.
}
\description{
Do numerical summaries by groups with formaula interface. Missing values are automatically removed.
}
\examples{
library(biostat)
data(cabbages, package = "MASS")

do_summary(~VitC, data = cabbages) \%>\%
    print(digits = 2)

do_summary(VitC ~ Cult, data = cabbages) \%>\%
    print(digits = 2)

do_summary(VitC ~ Cult + Date, data = cabbages, stat = "mean") \%>\%
    print(digits = 2)

do_summary(HeadWt + VitC ~ Cult + Date,
           data = cabbages,
           stat = c("n", "mean", "sd")) \%>\%
    print(digits = 1)

}
