% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.R
\docType{data}
\name{format_p_values}
\alias{format_p_values}
\alias{format_p_values.default}
\alias{format_p}
\alias{format_p_values.data.frame}
\alias{add_signif_stars}
\alias{get_signif_stars}
\alias{signif_syms_05s}
\alias{signif_syms_01s}
\alias{signif_syms_001s}
\alias{signif_syms_001}
\alias{signif_syms_01}
\alias{signif_syms_05}
\alias{signif_syms}
\alias{signif_stars_legend}
\alias{signif_stars_legend_2}
\alias{rm_zero}
\alias{format_as_p_columns}
\title{[!!!] Format p-values}
\format{An object of class \code{numeric} of length 1.}
\usage{
format_p_values(p, digits_p = 3, ..., alpha = 0.05, signif_stars = TRUE,
  rm_zero = FALSE, add_p = FALSE, rm_spaces = FALSE, ss = signif_syms)

\method{format_p_values}{default}(p, digits_p = 3, ..., signif_stars = TRUE,
  rm_zero = FALSE, add_p = FALSE, rm_spaces = FALSE, ss = signif_syms)

format_p(p_i, digits_p = 3, signif_stars = TRUE, rm_zero = FALSE,
  add_p = FALSE, rm_spaces = FALSE, ss = signif_syms)

\method{format_p_values}{data.frame}(data, digits_p = 3, cols = NULL, ...,
  ss = signif_syms, signif_stars = TRUE, rm_zero = FALSE, add_p = FALSE,
  rm_spaces = FALSE)

add_signif_stars(p, ss = signif_syms)

get_signif_stars(p, ss = signif_syms)

signif_syms_05s

signif_syms_01s

signif_syms_001s

signif_syms_001

signif_syms_01

signif_syms_05

signif_syms

signif_stars_legend(ss = signif_syms)

signif_stars_legend_2(ss = signif_syms, decreasing = FALSE)

rm_zero(str, dec = ".")

format_as_p_columns(data, colnames = c("p.value", "p.adjust"), digits_p = 3,
  rm_zero = FALSE, signif_stars = FALSE, ...)
}
\arguments{
\item{p}{A (vector of) p-value(s). Numeric or coercible to numeric.}

\item{digits_p}{(numeric) Number of significant digits to round a p value to.
                  No less than 2. \itemize{
\item if \code{digits_p = 2}: \enumerate{
      \item values below 0.001 are printed as \code{"<0.001"};
      \item values between 0.001 and 0.01 as \code{"<0.01"};
      \item all other values are rounded to two decimal places.
      }
\item if \code{digits_p = 3}, only formatting \code{"<0.01"} is skipped.
}}

\item{...}{Arguments to further methods.}

\item{signif_stars}{(logical) Flag if significance stars should be added to each p value.
Not less than 2.}

\item{rm_zero}{(logical) Flag if leading zero before the point should be
removed.}

\item{add_p}{(logical) Flag if letter "p" should included in the expression.}

\item{rm_spaces}{(logical) Flag if all spaces should be removed.}

\item{p_i}{(numeric) A single p value.}

\item{data}{A dataset.}

\item{str}{A string or convertible to string}

\item{colnames}{(character) vector with column names to be formatted as p values.}
}
\value{
A character vector with formatted p values.
}
\description{
Fuctions to fromat p values.
}
\details{
\itemize{
 \item \code{format_p} - formats single p value.
 \item \code{format_as_p_columns} - formats indicated numeric columns in a dataframe as p values (columns are converted into strings).
 \item \code{get_signif_stars} -  takes numeric p values brings appopriate stars of statistical significance.
 \item \code{add_signif_stars} - formats numeric p values by adding significance stars (result is character vector).
 \item \code{signif_stars_legend} - generates legend for significance stars (result is a string).
 \item \code{rm_zero} - function removes zero at the beginning of a number (returns a  string with the same value but without the leading zero).
}
}
\examples{
library(biostat)

# Prettify p-values

format_p_values(0.0005)

format_p_values(0.005)
format_p_values(0.005, signif_stars = FALSE)
format_p_values(0.005, rm_zero = TRUE)
format_p_values(0.005, digits_p = 2)
format_p_values(0.005, digits_p = 2, rm_zero = TRUE, signif_stars = FALSE)

format_p_values(0.00022)
format_p_values("0.00022")
format_p_values("0.052")

format_p_values(c(0.005, 0.0005, 0.052147))


get_signif_stars(0.005)
add_signif_stars(0.005)

get_signif_stars(0.0005)
add_signif_stars(0.0005)

get_signif_stars(0.052147)
add_signif_stars(0.052147)

signif_stars_legend()


format_p(0)
format_p(.02, digits_p = 2)
format_p(.0002)
format_p(.0002, signif_stars = FALSE)
format_p(.0002, ss = c("*****" = 0.001))
rm_zero(0.020)

}
\keyword{datasets}
